requirements.txt
requests>=2.25.0
websocket-client>=1.2.0
flask>=2.0.0 